% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_locations.R
\name{plot_current_users}
\alias{plot_current_users}
\title{Plots the Locations of Users in a Specified Room}
\usage{
plot_current_users(room_id = "25312", type = "leaflet")
}
\arguments{
\item{room_id}{The ID number of the room you want to retrieve. Defaults to
\code{"25312"} which is the "R Public" room.}

\item{type}{Can be either \code{"leaflet"} or \code{"rworldmap"}. Defaults to
\code{"leaflet"}.}
}
\value{
A map.
}
\description{
Scrapes the users in a specified Stack Overflow chat room, uses the Stack
Overflow API to geocode the locations, and plots the locations on a map using
either the "leaflet" or "rworldmap" packages.
}
\examples{

plot_current_users()
plot_current_users(type = "rworldmap")
plot_current_users("106")

}

