% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/room_rep.R
\name{room_rep}
\alias{room_rep}
\title{Sum of the Reputation of Current Room Occupants}
\usage{
room_rep(room_id = "25312")
}
\arguments{
\item{room_id}{The ID number of the room you want to retrieve. Defaults to
\code{"25312"} which is the "R Public" room.}
}
\value{
A single integer.
}
\description{
Sums the reputation of users currently in the specified room.
}
\examples{

room_rep()
room_rep("106")

}

