% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_locations.R
\name{user_list}
\alias{user_list}
\title{Lists the Locations of Users in a Specified Room}
\usage{
user_list(room_id = "25312")
}
\arguments{
\item{room_id}{The ID number of the room you want to retrieve. Defaults to
\code{"25312"} which is the "R Public" room.}
}
\value{
A \code{data.table} with the following columns: "reputation", 
"user_id", "location", "display_name", "lon", and "lat".
}
\description{
Scrapes the users in a specified Stack Overflow chat room and uses the Stack
Overflow API to geocode the locations, when available.
}
\examples{

user_list()
user_list("106")

}

