% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_locations.R
\name{user_location}
\alias{user_location}
\title{Lists the Locations of Specified Users}
\usage{
user_location(user_ids)
}
\arguments{
\item{user_ids}{The ID numbers of the users you want to retrieve the 
locations of. Can be a character or numeric vector.}
}
\value{
A \code{data.frame} with the following columns: "reputation", 
"user_id", "location", and "display_name".
}
\description{
Uses the Stack Overflow API to list the locations of specified users.
}
\examples{

user_location(1270695)
user_location(c(324364, 1270695))

}

